Moved AdvancedBan integration into core
to prevent code duplication
