+++ Bungee/build.gradle
@@ -39,8 +39,6 @@ ext {
 dependencies {
     compile 'net.md-5:bungeecord-api:1.10-SNAPSHOT'
     compile 'net.minecrell.mcstats:statslite-bungee:0.2.2'
-    
-    provided name: 'AdvancedBan-2.1.4-RELEASE'
 }
 
 processResources {
