+++ Bungee/src/main/java/net/minecrell/serverlistplus/bungee/BungeePlugin.java
@@ -53,7 +53,6 @@ import net.md_5.bungee.api.plugin.Listener;
 import net.md_5.bungee.api.plugin.TabExecutor;
 import net.md_5.bungee.event.EventHandler;
 import net.minecrell.mcstats.BungeeStatsLite;
-import net.minecrell.serverlistplus.bungee.integration.AdvancedBanBanDetector;
 import net.minecrell.serverlistplus.core.ServerListPlusCore;
 import net.minecrell.serverlistplus.core.ServerListPlusException;
 import net.minecrell.serverlistplus.core.config.PluginConf;
@@ -64,6 +63,7 @@ import net.minecrell.serverlistplus.core.logging.JavaServerListPlusLogger;
 import net.minecrell.serverlistplus.core.logging.ServerListPlusLogger;
 import net.minecrell.serverlistplus.core.player.ban.BanDetector;
 import net.minecrell.serverlistplus.core.player.ban.NoBanDetector;
+import net.minecrell.serverlistplus.core.player.ban.integration.AdvancedBanBanDetector;
 import net.minecrell.serverlistplus.core.plugin.ScheduledTask;
 import net.minecrell.serverlistplus.core.plugin.ServerListPlusPlugin;
 import net.minecrell.serverlistplus.core.plugin.ServerType;
